Move access control related tests to appropriate class
